pyproject.toml
[project]
name = "django-rest-envelope"
version = "0.1.0"
description = "A mixin for Django REST package List Views to return an object instead of a list."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    {name = "Ferran Jovell", email = "ferran.jovell+gh@gmail.com"}
]
keywords = ["django"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
    "Environment :: Web Environment",
    "Framework :: Django"
]
 dependencies = [
    "Django>=3.2.14",
    "djangorestframework>=3.11.0",
 ]

[project.urls]
homepage = "https://github.com/mrswats/django-rest-envelope"
"Bug Tracker" = "https://github.com/mrswats/django-rest-envelope/issues"

[project.optional-dependencies]
test = [
    "pre-commit",
    "pytest",
    "pytest-django",
]

[build-system]
requires = ["setuptools>=67.4"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["rest_envelope"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.black]
line-length = 99
target_version = ["py38"]
skip-string-normalization = false
include = "\\.pyi?$"
exclude = """
(
  /(
      \\.eggs         # exclude a few common directories in the
    | \\.git          # root of the project
    | \\.hg
    | \\.mypy_cache
    | \\.tox
    | \\.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
"""
